minor openApi changes
Look for both .json and .yaml files.

If neither is found, do not log as an error with a stack trace, since
"nothing broke".

targetUrl for generating proxy is actually a dummy target, since the
real target is in the CF header. But that URL is also used to fetch the
OpenAPI files; if it is https, the call will fail in local pcfdev with
its self-signed cert. So switch dummy target to http.