Get key and secret (and org/env) from VCAP_SERVICES
Instead of scraping key and secret from org-env-cache-config.yaml file,
get credentials from a bound service.

For now, this would be through a user-provided service instance, like:

  cf cups edgemicro-creds -p "org,env,key,secret"

which will prompt for the values. Key and secret are still generated
through edgemicro configure command; then copy/pasted.

The UPSi is then bound to the edgemicro app in the usual way. The
.edgemicro directory still needs to be copied into the app, since all the
other configuration is there.